quaternion switch implemented
It was necessary to replace trackingdata_to_matrices(data_EMT,
'CppCodeQuat') in a lot of files! Quaternions have to be read like 1:3
in TrackingFusion Files and like 2:4 in NDI files. really a pain!
Also, if you run the last section of wip_Felix, you can watch funny
Videos of paths that we went in 3D.