implemented mean_transformation.m
Attended the lecture today and realized that we need a new way to
calculate a mean transformation. This way ist really awesome and should
always be correct. Will incorporate it in all our files. If it should
not be correct, I commit now so we can revert the other files later.
Will try to -tag- this commit, a nice GitHub functionality I read about
yesterday. You can mark special commits with a name or phrase to find it
better when you need it.